Switch to relative image comparisons
Previously, the tolerence argument would count a pixel change as a fully different pixel. If the antialiasing changed just slightly from one image to another, each pixel along the alias would be counted as not matching, even if it was just a shade off of the original.

This changes the comaprison algorithm to take into account how different each pixel is. The significally changes the weight of the tolerance parameter, but will allow small changes to an image that effect a large number of pixels will be weighted similarly to a small number of pixels that were completely changed.